Added PDF generation feature to RenderImage demo (#5)
* RenderImage - Updated XF to latest 3.4 version

* RenderImage - Updated other dependencies

Updated UWP NetCore to 6.2.8, updated Xamarin.Essentials from preview to v1.1 release

* RenderImage - Updated Telerik Xamarin to 2019.2.318

* Finished updating Android project to SDK 28

* RenderImage - Finished upgrades (Android cropping confirmed working)

Remoevd unused Xamarin.Essentials and tested on platforms.

* RenderImage - Added proportional rendering to IRenderService.

This new feature allows the crop rectangle to be scaled up to the final display using a proportional percentage of the XF scaling

* RenderImage - Major Update: Added proportional crop with demo page

The MainPage will redner the content and send it to a details page. ProportionalCrop lets you pass a percentage value proportional to the full resolution size.

* RenderImage - Added PDF generation feature

The demo will now upload th eimage to the cloud to generate a PDF document. That document is returned and displayed in a RadPdfViewer.

* RenderImage - Updated readme with the new feature information